feat : token payload dto, TokenPayload custom param, users logged info inside each request

+++ backend/src/user/user.service.ts
@@ -3,6 +3,7 @@ import {
   HttpStatus,
   Injectable,
   NotFoundException,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
@@ -10,6 +11,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
 import { Repository } from 'typeorm';
 import { HashingService } from 'src/auth/hashing/hashing.service';
+import { tokenPayloadDto } from 'src/auth/dto/token-payload.dto';
 
 @Injectable()
 export class UserService {
@@ -43,17 +45,18 @@ export class UserService {
     return users;
   }
 
-  async findOne(id: number) {
+  async findOne(id: number, tokenPayload: tokenPayloadDto) {
     const user = await this.userService.findOne({
       where: {
         id,
       },
     });
     if (!user) throw new NotFoundException('User not found');
+    if(tokenPayload.sub!=id) throw new UnauthorizedException('Cannot access information about this user')
     return user;
   }
 
-  async update(id: number, updateUserDto: UpdateUserDto) {
+  async update(id: number, updateUserDto: UpdateUserDto, tokenPayload: tokenPayloadDto) {
     const partialUser = {
       name: updateUserDto?.name,
     };
@@ -68,17 +71,21 @@ export class UserService {
     });
 
     if (!personToUpdate) throw new NotFoundException('User not found');
+    if(tokenPayload.sub!=id) throw new UnauthorizedException('Cannot update information about this user')
 
     return this.userService.save(personToUpdate);
   }
 
-  async remove(id: number) {
+  async remove(id: number, tokenPayload: tokenPayloadDto) {
     const user = await this.userService.findOne({
       where: {
         id,
       },
     });
+
     if (!user) throw new NotFoundException('User not found');
+    if(tokenPayload.sub!=id) throw new UnauthorizedException('Cannot remove this user');
+
     return this.userService.remove(user);
   }
 }
